+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support for GHC 9.4.*
 * Values are pretty printed using `pretty-simple`
 * Timeouts are suspended during breakpoints for GHC >= 9.2
+* Fix a bug with monadic binds in do blocks
 
 ## 0.1.0.0 -- YYYY-mm-dd