+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Timeouts are suspended during breakpoints for GHC >= 9.2
 * Fix a bug with monadic binds in do blocks
 
+TODO: Add function that takes list of var names to ignore
+
 ## 0.1.0.0 -- YYYY-mm-dd
 
 * First version. Released on an unsuspecting world.